pyproject.toml
[project]
name = "bedrock-rag-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.243.0",
    "constructs>=10.5.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
]
